Adicionando funcionalidades da aula 4 - Ainda nao finalizada

+++ Desenvolva aplicações Web com JavaScript/03-JS para Web - armazenando dados no navegador/js/main.js
@@ -2,24 +2,29 @@ const form = document.getElementById("novoItem");
 const lista = document.getElementById("lista");
 const listaItens = JSON.parse(localStorage.getItem("itens")) || [];
 
+console.log(listaItens);
+
 listaItens.forEach((element) => {
-  criaElemento(element.nome, element.quantidade);
+  criaElemento(element);
 });
 
 form.addEventListener("submit", (evento) => {
   evento.preventDefault();
-
   const nome = evento.target.elements["nome"];
   const quantidade = evento.target.elements["quantidade"];
 
-  adicionaItem(nome.value, quantidade.value);
+  const item = {
+    nome: nome,
+    quantidade: quantidade,
+  };
+  adicionaItem(item);
 
-  nome.value = "";
-  quantidade.value = "";
+  item.nome = "";
+  item.quantidade = "";
 });
 
-function criaElemento(nome, quantidade) {
-  if (nome === "" || quantidade === "") {
+function criaElemento(item) {
+  if (item.nome === "" || item.quantidade === "") {
   } else {
     const novoItem = document.createElement("li");
     novoItem.classList.add("item");
@@ -29,26 +34,29 @@ function criaElemento(nome, quantidade) {
 
     novoItem.appendChild(quantidadeItem);
     novoItem.innerHTML += nome;
+    novoItem.dataset.id = item.id;
+    console.log(novoItem.querySelector("[data-id]"));
 
     lista.appendChild(novoItem);
   }
 }
 
-function adicionaItem(nome, quantidade) {
-  itemAtual = {
-    nome: nome,
-    quantidade: quantidade,
+function adicionaItem(item) {
+  if (verificaExistente(item)) {
+  } else {
+    criaElemento(item);
 
-    // functions: {
-    //   getNome() {
-    //     return this.nome;
-    //   },
-    // },
-  };
+    listaItens.push(item);
 
-  criaElemento(nome, quantidade);
+    localStorage.setItem("itens", JSON.stringify(listaItens));
+  }
+}
 
-  listaItens.push(itemAtual);
-
-  localStorage.setItem("itens", JSON.stringify(listaItens));
+function verificaExistente(item) {
+  const existe = listaItens.find((elemento) => elemento.nome === item.nome);
+  if (existe) {
+    return true;
+  } else {
+    return false;
+  }
 }
